DOCMAKE: Able to evaluate/plot 2D rates:k=k(x,y)
Generalizing the evaluation function was quite a mess...
It works (until you find a bug), but I suggest to rewrite the whole
evaluate/extrapolate/plot section of the rates, keeping in mind
to generalise this with rate functions and multi-D rates.
The differences/use of the valsRange is not clear and there
probably is some overhead now.